requirements.txt
flask==2.1.2
flask-sqlalchemy==2.5.1
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.28.0
redis==4.3.4
email-validator==1.2.1
flask-jwt-extended==4.4.2
flask-bcrypt==1.0.1
python-dotenv==0.20.0
psycopg2-binary==2.9.3
gevent==21.12.0
gunicorn==20.1.0
pytest==7.1.2
alembic==1.8.0
flasgger==0.9.5
